Stop Codon Pseudocode
Just a crappy initial outline so that people know where I'm thinking about starting with this. I'm not planning on leaving it as is (can definitely see a lot of areas already that could be significantly improved), but this gives a general idea of the rules I thought most important. Also, please don't be shy about correcting me - I may have noticed the mistake already, but it's good to know there are other other eyes on my work.